Prevent incorrect patterns from being added to the `deny` list.
